Index

OWENS.Bin

OWENS.Blade

OWENS.BladeData

OWENS.CactusGeom

OWENS.DriveShaftProps

OWENS.Inputs

OWENS.Iteration_Parameters

OWENS.NuMad

OWENS.Strut

OWENS.plyproperties

OWENS.Unsteady

OWENS.Unsteady_Land

OWENS.calcHydroResidual

OWENS.calculateDriveShaftReactionTorque

OWENS.calculateElementOrientation

OWENS.calculatePsiTheta

OWENS.count_cycles

OWENS.createGeneralTransformationMatrix

OWENS.createInitCondArray

OWENS.createSingleRotationDCM

OWENS.create_arcus_mesh

OWENS.create_hawt_biwing_mesh

OWENS.create_hawt_mesh

OWENS.create_mesh_struts

OWENS.externalForcing

OWENS.extrap_pred_vals

OWENS.find_boundary_vals

OWENS.find_range

OWENS.frame_convert

OWENS.generateOutputFilename

OWENS.getGustVel

OWENS.getOWENSPreCompOutput

OWENS.getRotorPosSpeedAccelAtTime

OWENS.getSectPropsFromOWENSPreComp

OWENS.mapACDMS

OWENS.mapAD15

OWENS.mapCactusLoadsFile

OWENS.my_getABD!

OWENS.my_getplystrain

OWENS.my_getz

OWENS.mywrite_vtk

OWENS.mywrite_vtk

OWENS.omegaSpecCheck

OWENS.owens

OWENS.plotMesh

OWENS.rainflow

OWENS.readBCdata

OWENS.readBladeData

OWENS.readCactusGeom

OWENS.readElementData

OWENS.readGeneratorProps

OWENS.readMesh

OWENS.readNuMadGeomCSV

OWENS.readNuMadMaterialsCSV

OWENS.readResultsModalOut

OWENS.rigidBodyRotation

OWENS.run34m

OWENS.runDLC

OWENS.simpleGenerator

OWENS.sort_peaks

OWENS.sum_cycles

OWENS.timeIntegrateSubSystem

OWENS.transMat

OWENS.trapezoidalRule

OWENS.updateRotorRotation

OWENS.userDefinedRotorSpeedProfile

OWENS.viz

OWENS.writeOwensNDL

Types and functions

OWENS.Bin — Type

Inputs pointing to the file paths of compiled binaries of external libraries

OWENS.Blade — Type

Internal, struct containing the CACTUS geometry file data for a blade

OWENS.BladeData — Type

Internal, struct containing blade specific data and location within the mesh

OWENS.CactusGeom — Type

Internal, struct containing the CACTUS geometry file data

OWENS.DriveShaftProps — Type

Internal, driveshaft stiffness k and damping c

OWENS.Inputs — Method

Inputs(;analysisType = "TNB", turbineStartup = 0, usingRotorSpeedFunction = false, tocp = [0.0,1.1], tocpVinf = [0.0,1.1], numTS = 50.0, deltat = 2e-3, Omegaocp = [7.2,7.2] ./ 60, Vinfocp = [12.0,12.0], aeroLoadsOn = 1, AD15On = false, driveTrainOn = false, generatorOn = false, hydroOn = false, topsideOn = true, interpOrder = 2, hdinputfile = "none", mdinputfile = "none", JgearBox = 0.0, gearRatio = 1.0, gearBoxEfficiency = 1.0, useGeneratorFunction = false, generatorProps = 0.0, ratedTorque = 0.0, zeroTorqueGenSpeed = 0.0, pulloutRatio = 0.0, ratedGenSlipPerc = 0.0, OmegaGenStart = 0.0, omegaControl = false, OmegaInit = 7.2/60, #TODO: simplify this in the code since it is redundant rigid = false, #turn off structural dynamics aeroloadfile = "modulepath/../test/data/inputfilestest/DVAWT2BLCDTElementData.csv", owensfile = "modulepath/../test/data/inputfilestest/15mTowertransientdvawtc2_lcdt.owens", outFilename = "none", numDofPerNode = 6, bladeData = [], driveShaftProps = DriveShaftProps(0.0,0.0) TOl = 1e-4, MAXITER = 300, iterwarnings = true, )

Model inputs for OWENS coupled analysis, struct

Inputs

analysisType::string: Newmark Beta time stepping "TNB", Dean time stepping "TD", modal "M"

turbineStartup::int: 1 forced start-up using generator as motor, 2 self-starting mode, 0 specified rotor speed mode")

usingRotorSpeedFunction::bool: use user specified rotor speed profile function

tocp::Array{<:float}: = time points for rotor speed profile (s)

tocp_Vinf::Array{<:float}: = time points for specified Vinf profile (s)

numTS::int: total number of timesteps to run

delta_t::float: timestep interval (s)

Omegaocp::Array{<:float}: = rotor speed points for rotor speed profile (Hz)

Vinfocp::Array{<:float}: = rotor speed points for specified Vinf profile (Hz)

aeroLoadsOn::bool: #0 off, 1 one way, 1.5 one way with deformation from last timestep, 2 two way

AD15On::bool: flag to use AD15 for aero

driveTrainOn::bool: flag to include drivetrain effects

generatorOn::bool: flag to include generator effects

hydroOn::bool: flag to include platform coupling

interpOrder::int: order used for extrapolating inputs and states, 0 flat, 1 linear, 2 quadratic

hd_input_file::string: file path to the HydroDyn .dat input file

md_input_file::string: file path to the MoorDyn .dat input file

JgearBox::float: gearbox intertia, standard SI units

gearRatio::float: gearbox gear ratio

gearBoxEfficiency::float: gearbox efficiency (typically 0-1)

useGeneratorFunction::bool: = flag to use user specified generator profile

generatorProps::float: not used, should clean up

ratedTorque::float: Generator rated max torque

zeroTorqueGenSpeed::float: rated generator speed (minus slippage)

pulloutRatio::float: Fraction of the min/max torque that the generator engages/disengages

ratedGenSlipPerc::float: extra speed from slipping?

OmegaGenStart::float: speed (Hz) at which generator would kick in

omegaControl::bool: false for fixed speed, true for dynamic

OmegaInit::float: initial rotor speed (Hz)

aeroloadfile::string: string of the name and path for the cactus aeroloads if using the old serial owens call

owensfile::string: string of the name and path for the owens input file if using the old serial owens call

potflowfile::string: string of the prefix and path for the directory containing the potential flow files from WAMIT (required by HydroDyn)

outFilename::string: path and name of output file, will be overwritten if already exists

numDofPerNode::int: number of degrees of freedom per node

bladeData::BladeData: see ?BladeData, only used if calling the old serial owens function

driveShaftProps::DriveShaftProps: see ?DriveShaftProps

TOl::float: gauss-seidel iteration tolerance

MAXITER::int: gauss-seidel maximum iterations

iterwarnings::bool: iteration warnings flag

Outputs:

OWENS.Inputs:

OWENS.Iteration_Parameters — Type

Internal, gauss-seidel iteration parameters

OWENS.NuMad — Type

NuMad(nweb,nstack,nsegments,span,airfoil,tetype,twistd,chord,xoffset,aerocenter,stackmattypes,stacklayers,segments,DPtypes,skin,webstack,webdp)

Parameters defining the blade composite layup. See NuMad user guide SAND2012_7028 appendix B for more details

Arguments

n_web::Int64: number of shear webs

n_stack::Int64: number of predefined composite stacks

n_segments::Int64: number of segments around the airfoil

span::Vector{Float64}: span-wise position

airfoil::Vector{String}: airfoil name

te_type::Vector{String}: trailing edge type

twist_d::Vector{Float64}: twist_d in degrees

chord::Vector{Float64}: chord length

xoffset::Vector{Float64}: The distance from the “nose” of a station to the blade reference axis.

aerocenter::Vector{Float64}: This is an aerodynamic parameter that is an output from aerodynamic performance analysis of a two-dimensional airfoil section. The aerodynamic center is the point along the chord where the aerodynamic pitching moment does not vary with changes in angle of attack.

stack_mat_types::Vector{Int64}: Material numbers used that correspond to each stack number

stack_layers::Array{Int64,2}: number of layers at each span used corresponding to each material type (first index corresponds to spanwise position, second index corresponds to the stack number)

segments::Array{Float64,2}: normalized starting and stopping points of each section (i.e. leading edge, sparcap, etc). First index corresponds to spanwise position, second index corresponds to the section, except there is an extra first column starting at -1 for the trailing edge. There must be a leading edge position at 0, and the last column must be 1 corresponding to the trailing edge again. Positions are fractions of the chord, lower (HP) is negative, upper (LP) is positive

DPtypes::Array{Int64,2}: division point types (NOTE THAT THIS ISN'T IMPLEMENTED AND DOES NOTHING CURRENTLY, i.e. only SINGLE is being used). First index corresponds to spanwise positoin, second corresponds to section number

skin_seq::Array{Seq,2}: stack sequence, is an array of structures, each containing a Vector{Int64} of the sequence (i.e. skin[2,5].seq). First index corresponds to spanwise positoin, second index the section

web_seq::Array{Seq,2}: same format and meaning as skin sequence, but for the webs with the second index corresponding to the web number

web_dp::Array{Seq,2}: same format as skin sequence, but this corresponds to the section numbers the web connects to at the top and bottom at both edges. There are always four entries in the CSV list and the order goes as follows: inboard LP, inboard HP, outboard HP, outboard LP.

OWENS.Strut — Type

Internal, struct containing the CACTUS geometry file data for a strut

OWENS.plyproperties — Method

Struct containing

material names

Composites.Material structs for each material name - see ?Composites.Material

OWENS.Unsteady — Method

Unsteady(model,topModel,mesh,el,aero;getLinearizedMatrices=false)

Executable function for transient analysis. Provides the interface of various external module with transient structural dynamics analysis capability.

# Input
* `inputs::Inputs`: see ?Inputs
* `topModel::FEAModel`: see ?OWENSFEA.FEAModel
* `mesh::Mesh`: see ?OWENSFEA.Mesh
* `el::El`: see ?OWENSFEA.El
* `bin::Bin`: see ?Bin
* `aero::function`: Fexternal, Fdof = aero(t) where Fexternal is the force on each affected mesh dof and Fdof is the corresponding DOFs affected
* `getLinearizedMatrices::Bool`: Flag to save the linearized matrices
* `elStorage::ElStorage.ElStorage`: Optional object containing stored element matrices
* `u_s::Array{<:float}`: Optional warm start of top deflections, of length Nnodes x Ndof

# Output
* `t`: time array
* `aziHist`: azimuthal history array
* `OmegaHist`: rotational speed array history
* `OmegaDotHist`: rotational acceleration array history
* `gbHist`: gearbox position history array
* `gbDotHist`: gearbox velocity history array
* `gbDotDotHist`: gearbox acceleration history array
* `FReactionHist`: Nodal reaction 6dof forces history
* `rigidDof`:
* `genTorque`: generator torque history
* `genPower`: generator power history
* `torqueDriveShaft`: driveshaft torque history
* `uHist`: mesh displacement history for each dof
* `epsilon_x_hist`: strain history for eps_xx_0 for each dof
* `epsilon_y_hist`: strain history for eps_xx_z for each dof
* `epsilon_z_hist`: strain history for eps_xx_y for each dof
* `kappa_x_hist`: strain history for gam_xz_0 for each dof
* `kappa_y_hist`: strain history for gam_xz_y for each dof
* `kappa_z_hist`: strain history for gam_xy_0 for each dof

OWENS.Unsteady_Land — Method

Unsteady(model,topModel,mesh,el,aero;getLinearizedMatrices=false)

Executable function for transient analysis. Provides the interface of various external module with transient structural dynamics analysis capability.

# Input
* `inputs::Model`: see ?Model
* `topModel::FEAModel`: see ?OWENSFEA.FEAModel
* `mesh::Mesh`: see ?OWENSFEA.Mesh
* `el::El`: see ?OWENSFEA.El
* `bin::Bin`: see ?Bin
* `aero::function`: Fexternal, Fdof = aero(t) where Fexternal is the force on each affected mesh dof and Fdof is the corresponding DOFs affected
* `getLinearizedMatrices::Bool`: Flag to save the linearized matrices
* `elStorage::ElStorage.ElStorage`: Optional object containing stored element matrices
* `u_s::Array{<:float}`: Optional warm start of top deflections, of length Nnodes x Ndof

# Output
* `t`: time array
* `aziHist`: azimuthal history array
* `OmegaHist`: rotational speed array history
* `OmegaDotHist`: rotational acceleration array history
* `gbHist`: gearbox position history array
* `gbDotHist`: gearbox velocity history array
* `gbDotDotHist`: gearbox acceleration history array
* `FReactionHist`: Nodal reaction 6dof forces history
* `rigidDof`:
* `genTorque`: generator torque history
* `genPower`: generator power history
* `torqueDriveShaft`: driveshaft torque history
* `uHist`: mesh displacement history for each dof
* `epsilon_x_hist`: strain history for eps_xx_0 for each dof
* `epsilon_y_hist`: strain history for eps_xx_z for each dof
* `epsilon_z_hist`: strain history for eps_xx_y for each dof
* `kappa_x_hist`: strain history for gam_xz_0 for each dof
* `kappa_y_hist`: strain history for gam_xz_y for each dof
* `kappa_z_hist`: strain history for gam_xy_0 for each dof

OWENS.calcHydroResidual — Method

calc_hydro_residual(new_accels, new_hydro_frcs, md_frc, u, FMultiplier)

Internal, finds the residual of the platform accelerations/forces when adding in new values from HydroDyn, MoorDyn, and OWENSFEA.

Input

new_accels::Vector{<:float}: the new platform accelerations in the inertial reference frame from OWENSFEA (m/s/s)

new_hydro_frcs::Vector{<:float}: the new platform hydro forces in the inertial reference frame from HydroDyn (N)

md_frc::Vector{<:float}: the mooring forces at the platform in the inertial reference frame from MoorDyn (N)

u::Vector{<:float}: input vector containing platform loads and accelerations prior to the new additions

FMultiplier::Number: scaling factor used to make the loads/accelerations in the same magnitude for the Jacobians

Output

u_resid: vector containing residual between the input u and the new load/accelerations

OWENS.calculateDriveShaftReactionTorque — Method

calculateDriveShaftReactionTorque(driveShaftProps,thetaRotor,thetaGB,thetaDotRotor,thetaDotGB)

Internal, calculates reaction torque of driveshaft

#Input

driveShaftProps: object containing driveshaft properties

thetaRotor: azimuth position of rotor/rotor shaft (rad)

thetaGB: azimuth position of gearbox shaft (rad)

thetaDotRotor: angular velocity of rotor/rotor shaft (rad/s)

thetaDotGB: angular velocity of gearbox shaft (rad/s)

#Output

torque: reaction torque of drive shaft

OWENS.calculateElementOrientation — Method

calculateElementOrientation(mesh)

Calculates the orientation of elements in a mesh.

#Input

mesh::OWENSFEA.Mesh see ?OWENSFEA.Mesh object containing mesh data

#Output

elOr::OWENSFEA.Ort see ?OWENSFEA.Ort object containing element orientation data

OWENS.calculatePsiTheta — Method

calculatePsiTheta(v)

Calculates the orientation of a single element. A local element frame is related to a hub frame through a transformation matrix CHtoE (transforming a vector from an element frame E to a global frame H) such that CHtoE = [M2(Theta)]*[M3(Psi)]. Here [M2( )] is a direction cosine matrix about a 2 axis and [M3( )] is a direction cosine matrix about a 3 axis.

#Input

v: vector from node 1 to node 2 of an element

#Output

Psi: "3" angle for element orientation (deg)

Theta: "2" angle for element orientation (deg)

OWENS.count_cycles — Method

Count the cycles from the data

OWENS.createGeneralTransformationMatrix — Method

createGeneralTransformationMatrix(angleArray,axisArray)

Calculates the transformation matrix assocaited with a general Euler rotation sequence.

#Input

angleArray: = array of angles for Euler rotation sequence

axisArray: = array of axis of rotatoins for Euler rotation

#Output

dcmTotal: = transformation matrix of specified euler rotation sequence

OWENS.createInitCondArray — Method

setInitConditions(initDisps, numNodes, numDOFPerNode)

Creates the formatted initial conditions array needed by OWENSFEA

#Input

initDisps: an array of length numDOFPerNode specifying the initial displacement of each DOF

numNodes: the number of nodes in the given mesh

numDOFPerNode: the number of unconstrained degrees of freedom calculated in each node

#Output

initCond: array containing initial conditions. initCond(i,1) node number for init cond i. initCond(i,2) local DOF number for init cond i. initCond(i,3) value for init cond i.

OWENS.createSingleRotationDCM — Method

Creates a direction cosine matrix (dcm) associated with a rotation of angleDeg about axisNum.

OWENS.create_arcus_mesh — Method

createarcusmesh(;Ht = 15.0, Hb = 147.148-15.0, #blade height R = 54.014, # m bade radius nblade = 3, ntelem = 30, #tower elements nbelem = 30, #blade elements ncelem = 4, strutmountpoint = 0.01, # This puts struts at top and bottom bshapex = zeros(nbelem+1), #Blade shape, magnitude is irrelevant, scaled based on height and radius above bshapez = zeros(nbelem+1), jointtype = 0, cablesconnectedtobladebase = true, angularOffset = 0.0)

ARCUS mesh configuration: no tower between blades, no struts, but cables from top center attaching to specified blade mount point at base

#Inputs

Ht::float: height of tower before blades attach (m)

Hb::float: blade height (m)

R::float: bade radius (m)

nblade::int: number of blades

ntelem::int: number of tower elements

nbelem::int: number of blade elements

ncelem::int: number of strut elements

c_mount_ratio::float: factor of blade height where the struts attach on both top and bottom

bshapex::Array{<:float}: Blade shape, magnitude is irrelevant, scaled based on height and radius above

bshapez::Array{<:float}: Blade shape, magnitude is irrelevant, scaled based on height and radius above

joint_type::int: 0 is fixed, 1 is about x-axis, 2 is y-axis, etc

cables_connected_to_blade_base::bool: = true,

angularOffset::float: (rad) angular offset of mesh generation, typically used to match CACTUS input. Value of 0 puts blade 1 at the "north" position and the others populate counterclockwise when looking down

#Outputs

mymesh::OWENSFEA.Mesh: see ?OWENSFEA.Mesh

ort::OWENSFEA.Ort: see ?OWENSFEA.Ort

myjoint:Array{<:float}: see ?OWENSFEA.FEAModel.joint

OWENS.create_hawt_biwing_mesh — Method

createhawtbiwingmesh(;Ht = 15.0, Hb = 147.148-15.0, #blade height R = 54.014, # m bade radius nblade = 3, ntelem = 30, #tower elements nbelem = 30, #blade elements ncelem = 4, strutmountpoint = 0.01, # This puts struts at top and bottom bshapex = zeros(nbelem+1), #Blade shape, magnitude is irrelevant, scaled based on height and radius above bshapez = zeros(nbelem+1), jointtype = 0, cablesconnectedtoblade_base = true, angularOffset = 0.0)

#Inputs

Ht::float: height of tower before blades attach (m)

Hb::float: blade height (m)

R::float: bade radius (m)

nblade::int: number of blades

ntelem::int: number of tower elements

nbelem::int: number of blade elements

ncelem::int: number of strut elements

c_mount_ratio::float: factor of blade height where the struts attach on both top and bottom

bshapex::Array{<:float}: Blade shape, magnitude is irrelevant, scaled based on height and radius above

bshapez::Array{<:float}: Blade shape, magnitude is irrelevant, scaled based on height and radius above

joint_type::int: 0 is fixed, 1 is about x-axis, 2 is y-axis, etc

cables_connected_to_blade_base::bool: = true,

angularOffset::float: (rad) angular offset of mesh generation, typically used to match CACTUS input. Value of 0 puts blade 1 at the "north" position and the others populate counterclockwise when looking down

#Outputs

mymesh::OWENSFEA.Mesh: see ?OWENSFEA.Mesh

ort::OWENSFEA.Ort: see ?OWENSFEA.Ort

myjoint:Array{<:float}: see ?OWENSFEA.FEAModel.joint

OWENS.create_hawt_mesh — Method

createhawtmesh(;Ht = 15.0, Hb = 147.148-15.0, #blade height R = 54.014, # m bade radius nblade = 3, ntelem = 30, #tower elements nbelem = 30, #blade elements ncelem = 4, strutmountpoint = 0.01, # This puts struts at top and bottom bshapex = zeros(nbelem+1), #Blade shape, magnitude is irrelevant, scaled based on height and radius above bshapez = zeros(nbelem+1), jointtype = 0, cablesconnectedtobladebase = true, angularOffset = 0.0)

ARCUS mesh configuration: no tower between blades, no struts, but cables from top center attaching to specified blade mount point at base

#Inputs

Ht::float: height of tower before blades attach (m)

Hb::float: blade height (m)

R::float: bade radius (m)

nblade::int: number of blades

ntelem::int: number of tower elements

nbelem::int: number of blade elements

ncelem::int: number of strut elements

c_mount_ratio::float: factor of blade height where the struts attach on both top and bottom

bshapex::Array{<:float}: Blade shape, magnitude is irrelevant, scaled based on height and radius above

bshapez::Array{<:float}: Blade shape, magnitude is irrelevant, scaled based on height and radius above

joint_type::int: 0 is fixed, 1 is about x-axis, 2 is y-axis, etc

cables_connected_to_blade_base::bool: = true,

angularOffset::float: (rad) angular offset of mesh generation, typically used to match CACTUS input. Value of 0 puts blade 1 at the "north" position and the others populate counterclockwise when looking down

#Outputs

mymesh::OWENSFEA.Mesh: see ?OWENSFEA.Mesh

ort::OWENSFEA.Ort: see ?OWENSFEA.Ort

myjoint:Array{<:float}: see ?OWENSFEA.FEAModel.joint

OWENS.create_mesh_struts — Method

return mymesh, myort, myjoint, AD15bldNdIdxRng, AD15bldElIdxRng = createmeshstruts(;Ht = 15.0, Hb = 147.148-15.0, #blade height R = 54.014, # m bade radius AD15hubR = 2.0, nblade = 3, ntelem = 30, #tower elements nbelem = 30, #blade elements nselem = 4, struttwrmountpointbot = 0.01, # This puts struts at top and bottom struttwrmountpointtop = 0.99, # This puts struts at top and bottom strutbldmountpointbot = 0.01, # This puts struts at top and bottom strutbldmountpointtop = 0.99, # This puts struts at top and bottom bshapex = zeros(nbelem+1), #Blade shape, magnitude is irrelevant, scaled based on height and radius above bshapez = zeros(nbelem+1), bshapey = zeros(nbelem+1), # but magnitude for this is relevant angularOffset = 0.0, #Blade shape, magnitude is irrelevant, scaled based on height and radius above AD15_ccw = false, verbosity=0, # 0 nothing, 1 basic, 2 lots: amount of printed information connectBldTips2Twr =true)

Standard Mesh Matching 5MW, 34m configurations

#Inputs

Ht::float: height of tower before blades attach (m)

Hb::float: blade height (m)

R::float: bade radius (m)

nblade::int: number of blades

ntelem::int: number of tower elements

nbelem::int: number of blade elements

nselem::int: number of strut elements

strut_twr_mountpointbot::float = 0.01, # factor of blade height where the bottom strut attaches on the tower

strut_twr_mountpointtop::float = 0.99, # factor of blade height where the top strut attaches on the tower

strut_bld_mountpointbot::float = 0.01, # factor of blade height where the bottom strut attaches on the blade

strut_bld_mountpointtop::float = 0.99, # factor of blade height where the top strut attaches on the blade

bshapex::Array{<:float}: Blade shape, magnitude is irrelevant, scaled based on height and radius above

bshapez::Array{<:float}: Blade shape, magnitude is irrelevant, scaled based on height and radius above

bshapey::Array{<:float}: Blade shape, magnitude IS relevant #TODO: resolve this

angularOffset::float: (rad) angular offset of mesh generation, typically used to match CACTUS input. Value of 0 puts blade 1 at the "north" position and the others populate counterclockwise when looking down

AD15_ccw::boolean: Use AD15 convention of VAWT counter-clockwise with blade root at top (blade points down)

AD15hubR::float: AD15 has a hub radius, so the struts do not go all the way to the center of the axis of rotation, while the structural mesh does.

verbosity::int: 0 nothing, 1 basic, 2 lots: amount of printed information

connectBldTips2Twr::book: True for Darrieus style, false for H-VAWT, but the blade shapes should be appropriate

#Outputs

mymesh::OWENSFEA.Mesh: see ?OWENSFEA.Mesh

myort::OWENSFEA.Ort: see ?OWENSFEA.Ort

myjoint:Array{<:float}: see ?OWENSFEA.FEAModel.joint

AD15bldNdIdxRng: indices for start and end of all blades for AD15 (includes struts). Note that strut start nodes may be inside the strut (strut connects to tower, AD15 blade connects to hub wich is a few nodes away from tower)

AD15bldElIdxRng: range of elements for start and end of all AD15 blades (includes struts)

OWENS.externalForcing — Method

externalForcing(time,timeArray,ForceValHist,ForceDof)

Internal, linear time interpolation on the input forces (ForceValHist) for each specified ForceDof

This function specifies external forcing for a transient analysis. Fexternal is a vector of loads and Fdof is a corresponding vector of degrees of freedom the concentrated loads in Fexternal correspond to. The input time allows for arbitrary time varying loads The global degree of freedom number corresponding with the local degree of freedom of a node may be calculated by: globalDOFNumber = (nodeNumber-1)*6 + localDOFnumber The localDOFnumber may range from 1 to 6 such that 1 corresponds to a force in "x direction" of the co-rotating hub frame. 2 and 3 corresponds to a force in the "y" and "z directions" respectively. 4, 5, and 6 correspond to a moment about the "x", "y", and "z" directions respectively.

#Input

time: Current time

timeArray: time associated with ForceValHist

ForceValHist: Forces for each time for each Dof

ForceDof: Dofs within ForceValHist

#Output

Fexternal: vector of external loads (forces/moments)

Fdof: vector of corresponding DOF numbers to apply loads to

OWENS.extrap_pred_vals — Method

extrap_pred_vals(curr_vals, ts, t_out, interp_order)

Internal, calculates predicted values at t_out based on previous values at earlier times.

Input

curr_vals::Vector{<:float}: input values

ts::Vector{<:float}: time values corresponding to curr_vals. Left to right, times go from earliest to most recent.

t_out::float: time for values to be extrapolated tocp

interp_order::int: order of the spline fit for the extrapolation, 0 flat, 1 linear, 2 quadratic

Output

pred_vals: extrapolated values at t_out

OWENS.find_boundary_vals — Method

Find the minimum and maximum mean value and maximum range from a vector of cycles

OWENS.find_range — Method

Returns the range index where the value is found

OWENS.frame_convert — Method

frame_convert(init_frame_vals, trans_mat)

Internal, transfers 6 DOFs element-wise to a new reference frame

Input

init_frame_vals::Vector{<:float}: Values in 6 degrees of freedom in the initial reference frame

trans_mat::Array{<:float}: Transformation matrix to the output reference frame

Output

out_frame_vals: Values in 6 degrees of freedom in the output reference frame

OWENS.generateOutputFilename — Method

Internal, generates an output file name depending on the analysis type

OWENS.getGustVel — Method

time::TF: in seconds

nominalVinf::TF: Nominal velocity used to calculate the IEC gust size (m/s)

R::TF: Turbine Radius (m)

G_amp::TF: IEC gust amplitude (m/s)

gustT::TF: IEC gust duration (s)

gustDelayT::TF: IEC gust delay time

OWENS.getOWENSPreCompOutput — Method

getOWENSPreCompOutput(numadIn;yscale=1.0,plyprops = plyproperties())

Takes numad formatted inputs for composite layup and material properties and runs them through OWENSPreComp

#Inputs

numadIn::NuMad: see ?NuMad inputs

yscale::float: airfoil thickness scaling

plyprops::plyproperties(): see ?plyproperties for input material properties

#Outputs

precompoutput::OWENSPreComp.Input: see ?OWENSPreComp.Input

precompinput::OWENSPreComp.properties: see ?OWENSPreComp.properties

OWENS.getRotorPosSpeedAccelAtTime — Method

getRotorPosSpeedAccelAtTime(t0,time,aziInit)

Uses the user defined function rotorSpeedProfile() to get the azimuth, speed, and acceleration of the rotor.

#Input

t0 time at which azimuth integration is beginning

time current time that position, velocity, and acceleration are being requested

aziInit initial rotor azimuth angle integration will begin at

#Output

rotorAzimuth azimuth position of rotor (rad) at time

rotorSpeed rotor speed (Hz) at time

rotorAcceleration rotor acceleration (Hz/s) at time

OWENS.getSectPropsFromOWENSPreComp — Method

getSectPropsFromOWENSPreComp(usedUnitSpan,numadIn,precompoutput;GX=false,precompinputs=nothing)

Arranges the precomp output into the sectional properties required by OWENSFEA

#Inputs

usedUnitSpan::Array{<:float}: Array specifying the relative (0-1) mesh z locations where the sectional properties are actually called for (interpolation)

numadIn::NuMAD: see ?NuMAD

precompoutput::OWENSPreComp.properties: see ?OWENSPreComp.properties

GX::bool: optional specifies if GX outputs should be output

precompinputs: optional

#Outputs

sectionPropsArray::SectionPropsArray: see ?OWENSFEA.SectionPropsArray, if !GX bool

stiff: if GX bool

mass: if GX bool

stiff, mass

OWENS.mapACDMS — Method

mapACDMS(t,mesh,el)

map OWENSAero forces to OWENS mesh dofs

Inputs

t::float: time at which to get the loads (can be called repeatedly at the same time or for large time gaps, will infill run as needed)

mesh::OWENSFEA.Mesh: see ?OWENSFEA.Mesh

mesh::OWENSFEA.El: see ?OWENSFEA.El

Outputs:

ForceValHist::Array(<:float): Force or moment (N, N-m) at the time corresponding to the time specified

ForceDof::Array(<:int): DOF numbers cooresponding to forces (i.e. mesh element 1 has dofs 1-6, 2 has dofs 7-12, etc)

OWENS.mapAD15 — Method

mapAD15(t,mesh)

map AD15 forces to OWENS mesh dofs

Inputs

t::float: time at which to get the loads (can be called repeatedly at the same time or for large time gaps, will infill run as needed)

mesh::OWENSFEA.Mesh: see ?OWENSFEA.Mesh

mesh::OWENSFEA.El: see ?OWENSFEA.El

Outputs:

ForceValHist::Array(<:float): Force or moment (N, N-m) at the time corresponding to the time specified

ForceDof::Array(<:int): DOF numbers cooresponding to forces (i.e. mesh element 1 has dofs 1-6, 2 has dofs 7-12, etc)

OWENS.mapCactusLoadsFile — Method

Internal, takes cactus loads and geometry and OWENS mesh and maps the loads to the blades' FEA dofs

OWENS.my_getABD! — Function

my_getABD(matid::AbstractArray{<:Integer,1}, nply::AbstractArray{<:Integer,1}, tply::AbstractArray{<:Real,1}, theta::AbstractArray{<:Real,1}, q::AbstractArray{<:AbstractArray{<:Real,2},1}), offset::Real

my_getABD(lam::Laminate, q::AbstractArray{<:AbstractArray{<:Real,2},1}), offset::Real

Returns A, B, and D matrices

Arguments:

matid::AbstractArray{<:Integer,1}: material id of each lamina

nply::AbstractArray{<:Integer,1}: number of plies in each lamina

tply::AbstractArray{<:Real,1}: thickness of a ply (m) in each lamina

theta::AbstractArray{<:Real,1}: orientation (deg) of each lamina

q::AbstractArray{<:AbstractArray{<:Real,2}}: Stiffness matrix of each lamina

offset::Real: Optional, used if neutral axis is not centered on the ply. If used, specify distance from neutral axis to upper surface (m).

OWENS.my_getplystrain — Function

my_getplystrain(nply::AbstractArray{<:Integer,1}, tply::AbstractArray{<:Real,1}, theta::AbstractArray{<:Real,1}, resultantstrain::AbstractArray{<:Real,1}, offset::Real)

my_getplystrain(lam::Laminate, resultantstrain::AbstractArray{<:Real,1}, offset::Real)

Optional Argument

offset::Real: Optional, used if neutral axis is not centered on the ply. If used, specify distance from neutral axis to upper surface (m).

Calculates strains in each ply aligned with principal material direction.

OWENS.my_getz — Function

getz(tply::AbstractArray{<:Real,1}, nply::AbstractArray{<:Integer,1}, offset::Real)

getz(lam::Laminate, offset::Real)

Returns a laminate's z-coordinates (coordinates of top and bottom of laminas) given the thickness of plies in each lamina and the number of plies in each lamina. Offset is optional, defaults to 0.0, is the distance from the neutal axis to the upper ply if the laminate is not centered about the neutral axis of a beam.

OWENS.mywrite_vtk — Method

mywrite_vtk(name, assembly::Assembly, [state::AssemblyState, ]λ::Number,
    eigenstate::AssemblyState; scaling=1.0, mode_scaling=1.0, cycles=1,
    steps=100)

Write a series of files corresponding to the elastic motion of the assembly about the deformed state encoded in state defined by the eigenvalue λ and the eigenvector encoded in eigenstate over the time period specified by time.

The steady-state deflections can be scaled with scaling and the eigenmode deflections can be scaled using mode_scaling.

The current time is encoded in the metadata tag "time"

OWENS.mywrite_vtk — Method

mywrite_vtk(name, assembly::Assembly; kwargs...)
mywrite_vtk(name, assembly::Assembly, state::AssemblyState; kwargs...)
mywrite_vtk(name, assembly::Assembly, history::Vector{<:AssemblyState}], dt;
    kwargs...)

Write the deformed geometry (and associated data) to a VTK file for visualization using ParaView.

The state argument may be omitted to write the original geometry to a VTK file without any associated data.

If the solution time history is provided, the time step must also be provided

Keyword Arguments

sections = nothing: Cross section geometry corresponding to each point, defined in a frame aligned with the body frame but centered around the corresponding point. Defined as an array with shape (3, ncross, np) where ncross is the number of points in each cross section and np is the number of points.

scaling=1.0: Parameter to scale the deflections (only valid if state is provided)

metadata=Dict(): Dictionary of metadata for the file(s)

OWENS.omegaSpecCheck — Method

Internal, gets specified rotor speed at time

OWENS.owens — Method

owens(owensfile,analysisType;
    delta_t=2e-3,
    numTS=100,
    tocp=[0.0,1.1],
    Omegaocp=[0.0,1.0],
    OmegaInit=0.0,
    OmegaGenStart=0.0,
    usingRotorSpeedFunction=false,
    nlOn=true,
    Omega=0.0,
    turbineStartup=0,
    spinUpOn=false,
    numModesToExtract=20,
    displInitGuess=0.0,
    airDensity=1.2041,
    aeroElasticOn = false,
    guessFreq = 0,
    gravityOn = true,
    generatorOn = false,
    omegaControl = false,
    iterationType = "NR", # nlParams
    adaptiveLoadSteppingFlag = true,
    tolerance = 1.0000e-06,
    maxIterations = 50,
    maxNumLoadSteps = 20,
    minLoadStepDelta = 0.0500,
    minLoadStep = 0.0500,
    prescribedLoadStep = 0.0,
    elementOrder = 1,
    numDofPerNode = 6,
    hydroOn = false,
    platformTurbineConnectionNodeNumber = 1,
    JgearBox =0.0,
    gearRatio = 1.0,
    gearBoxEfficiency = 1.0,
    useGeneratorFunction = false,
    generatorProps = 0.0,
    driveTrainOn = false)

Original serial and file reading method of running an analysis.

#Inputs See ?OWENS.Model and ?OWENSFEA.FEAModel

#Outputs See ?OWENS.Unsteady, ?OWENSFEA.Modal

OWENS.plotMesh — Method

plotMesh plots a mesh with various orthographic and isometric views plotMesh(mesh,colorstring,meshSeg)

This function plots a mesh in a 3D with isometric view.

#Input

mesh: object containing mesh data

colorstring: string containing color for plotting

meshSeg: array containing number of elements per mesh segment

ax1: handle to the 1 plot

ax2: handle to the 2 plot

ax3: handle to the 3 plot

ax4: handle to the 4 plot

#Output NONE

OWENS.rainflow — Method

rainflow(signal;nbins_range=10,nbins_mean=10)

Convenience function that returns the binned cycles with the corresponding ranges and means

Inputs

signal::Array{<:Real,1}: data input

nbins_range::Array{<:Int,1}: Number of bins for range

nbins_mean::Array{<:Int,1}: Number of bins for mean

m` : Wohler exponent (default is 3)

Teq` : The equivalent number of load cycles (default is 1, but normally the time duration in seconds is used)

Outputs:

Ncycles::Array{<:Real,2}: Summed/binned cycles with columns corresponding to mean levels and rows corresponding to range levels

meanIntervals::Array{<:Real,1}: Mean levels corresponging with bins columns

rangeIntervals::Array{<:Real,1}: Range levels corresponging with bins rows

equivalentLoad::Array{<:Real,1}: Design equivalent load for each mean level

OWENS.readBCdata — Method

readBCdata(bcfilename,numNodes,numDofPerNode)

This function reads the boundray condition file and stores data in the boundary condition object.

#Input

bcfilename::string: string containing boundary condition filename

numNodes::int: number of nodes in structural model

numDofPerNode::int: number of degrees of freedom per node

#Output

BC::OWENSFEA.BC_struct: see OWENSFEA.BC_struct, object containing boundary condition data

OWENS.readBladeData — Method

readBladeData(filename)

This function reads blade data from file

#Input

filename::string: string containing /path/to/bladedata.bld

#Output

bladeData::BladeData: see ?BladeData object containing blade data

OWENS.readCactusGeom — Method

Internal, reads cactus .geom file and stores each column in an array within the CactusGeom struct

OWENS.readElementData — Method

readElementData(numElements,elfile,ortfile,bldfile

Reads element data and stores data in the element data object.

#Input

numElements::int: number of elements in structural mesh

elfile::string: element data path/to/filename

ortfile::string: element orientation path/to/filename

bldfile::string: blade data path/to/filename

#Output

el::OWENSFEA.El: see OWENSFEA.El element data object

OWENS.readGeneratorProps — Method

readGeneratorProps(generatorfilename)

This function reads generator properties from file.

#Input

generatorfilenanme::string: = string containing path/to/generatorfile

#Output

genprops: = model object containing generator properties

OWENS.readMesh — Method

readMesh(filename)

Reads the mesh file and stores data in the mesh object.

input:

filename::string string containing mesh path/to/filename.mesh

output:

mesh::OWENSFEA.Mesh see OWENSFEA.Mesh

OWENS.readNuMadGeomCSV — Method

readNuMadGeomCSV(NuMadgeomxlscsv_file)

Parameters defining the rotor (apply to all sections).

Arguments

NuMad_geom_xlscsv_file::String: name of the numad excel CSV file being read (!!! THE NUMAD TAB MUST BE SAVED AS A CSV FOR THIS TO WORK !!!)

Returns

Output::NuMad: numad structure as defined in the NuMad structure docstrings.

OWENS.readNuMadMaterialsCSV — Method

readNuMadMaterialsCSV(NuMadmaterialsxlscsv_file)

Parameters defining the rotor materials.

Arguments

NuMad_materials_xlscsv_file::String: name of the numad excel CSV file being read (!!! THE NUMAD TAB MUST BE SAVED AS A CSV FOR THIS TO WORK !!!)

Returns

Output::plyproperties: plyproperties structure as defined in the plyproperties structure docstrings.

OWENS.readResultsModalOut — Method

Internal, reads modal file and returns freq, damp, and modeshapes

OWENS.rigidBodyRotation — Method

rigidBodyRotation(B1,B2,B3,AngleArray,AxisArray)

Performs a coordinate transformation from a local body "B"(element) frame to a common hub frame "H" via a 3-2-3 euler rotation sequence

#Input

B1: array containing body frame 1 coordinates of points to be mapped to the hub frame

B2: array containing body frame 2 coordinates of points to be mapped to the hub frame

B3: array containing body frame 3 coordinates of points to be mapped to the hub frame

AngleArray: Array of angles for Euler rotation sequence

AxisArray: Array of axes for Euler rotation sequence

#Output

H1: array containg hub frame 1 coordinates of points mapped to the hub frame from body frame

H2: array containg hub frame 2 coordinates of points mapped to the hub frame from body frame

H3: array containg hub frame 3 coordinates of points mapped to the hub frame from body frame

That is CHtoB = [M3(SweepAngle)][M2(Theta)][M3(Psi)];

OWENS.run34m — Method

run34m(inputs,feamodel,mymesh,myel,aeroForces,deformAero;steady=true,system=nothing,assembly=nothing,VTKFilename="./outvtk")

helper function that rearranges the outputs into the expected 34m output # Input * inputs::Model: see ?Model * topModel::FEAModel: see ?OWENSFEA.FEAModel * mesh::Mesh: see ?OWENSFEA.Mesh * el::El: see ?OWENSFEA.El * aeroForces::function: Fexternal, Fdof = aero(t) where Fexternal is the force on each affected mesh dof and Fdof is the corresponding DOFs affected * deformAero::function: see deformTurb(azi;newOmega=-1,newVinf=-1,bldx=-1,bldz=-1,bld_twist=-1,steady=false) * steady::bool: run steadystate with no aero or not * system: see ?GXBeam.System * assembly: see ?GXBeam.Assembly * VTKFilename::string: Unused: path and name of VTK output

# Output
* `eps_x`: strain history for eps_xx_0 for (Nbld,N_ts,mymesh.meshSeg[2]+1)
* `eps_y`: strain history for eps_xx_z like above
* `eps_z`: strain history for eps_xx_y like above
* `kappa_x`: strain history for gam_xz_0 like above
* `kappa_y`: strain history for gam_xz_y like above
* `kappa_z`: strain history for gam_xy_0 like above
* `t`: time array
* `FReactionHist`: Nodal reaction 6dof forces history
* `OmegaHist`: rotational speed array history
* `genTorque`: generator torque history
* `torqueDriveShaft`: driveshaft torque history
* `aziHist`: azimuthal history array
* `uHist`: mesh displacement history for each dof
* `epsilon_x_hist`: strain history for eps_xx_0 for each dof
* `epsilon_y_hist`: strain history for eps_xx_z for each dof
* `epsilon_z_hist`: strain history for eps_xx_y for each dof
* `kappa_x_hist`: strain history for gam_xz_0 for each dof
* `kappa_y_hist`: strain history for gam_xz_y for each dof
* `kappa_z_hist`: strain history for gam_xy_0 for each dof

OWENS.runDLC — Method

runDLC(DLCs,Inp,path; Vinfrange=LinRange(5,20,16), IECstd=""2"", WindChar=""A"", WindClass=1, turbsimpath="./turbsimfiles", templatefile="./templateTurbSim.inp", pathtoturbsim="../../openfast/build/modules/turbsim/turbsim", NumGridZ=100, NumGridY=100, Vref=10.0, Vdesign=11.0, grid_oversize=1.1, regenWindFiles=false)

# Input
* `DLCs`: ["1_1","1_2"]
* `Inp::MasterInput`: see ?OWENS.MasterInput
* `path`: desired path to run everything
* `Vinf_range`: =LinRange(5,20,16),
* `IEC_std`: =""2"",
* `WindChar`: =""A"",
* `WindClass`: =1,
* `turbsimpath`: ="./turbsimfiles", path where it dumps the turbsim files
* `templatefile`: ="./template_files/templateTurbSim.inp",
* `pathtoturbsim`: ="../../openfast/build/modules/turbsim/turbsim",
* `NumGrid_Z`: =100,
* `NumGrid_Y`: =100,
* `Vref`: =10.0,
* `Vdesign`: =11.0, # Design or rated speed
* `grid_oversize`: =1.1,
* `regenWindFiles`: =false

# Output
* `nothing`:

OWENS.simpleGenerator — Method

simpleGenerator(generatorProps,genSpeed)

Caclulates generator torque for simple induction generator

#Input

generatorProps object containing generator properties, see ?model

genSpeed::float generator speed (Hz)

#Output

genTorque::float generator torque

OWENS.sort_peaks — Function

This function sorts out points where the slope changes sign

OWENS.sum_cycles — Method

Sums the cycle count given intervals of rangeintervals and meanintervals. The rangeintervals and meanintervals are given in fraction of range size

OWENS.timeIntegrateSubSystem — Method

timeIntegrateSubSystem(M,K,C,F,delta_t,u,udot,uddot)

Internal, performs integration of a system using the Newmark-Beta method (constant-average acceleration sceheme).

#Input

M: system mass matrix

K: system sttiffness matrix

C: system damping matrix

F: system force vector

delta_t: time step

u: displacement at beginning of time step

udot: velocity at beginning of time step

uddot: acceleration at beginning of time step

#Output

unp1: displacement at end of time step

udotnp1: velocity at end of time step

uddotnp1: acceleration at end of time step

OWENS.transMat — Method

transMat(theta1, theta2, theta3)

Internal, computes the 3x3 transformation matrix for given input rotations. The generated matrix is the closest orthonormal matrix to the Bernoulli-Euler transformation matrix from beam theory, which assumes small rotations. A full description of this matrix is found in the "FASTCoordinateSystems.doc" document by Jason Jonkman.

OWENS.trapezoidalRule — Method

Internal, simple trapezoidal rule integration

OWENS.updateRotorRotation — Method

updateRotorRotation updates rotor rotation

updateRotorRotation(Irotor,Crotor,Krotor,shaftTorque,genTorque,azi_s,Omega_s,OmegaDot_s,delta_t)

Internal, updates the rotor rotation given rotor properties and external torques

#Input

Irotor: rotor inertia

Crotor: arbitrary rotor damping

Krotor: arbitrary rotor stiffness

shaftTorque: torque from external forces on rotor

genTorque: torque from generator

azi_s: rotor azimuth (rad) at beginning of time step

Omega_s: rotor speed (Hz) at beginning of time step

OmegaDot_s: rotor acceleration (Hz/s) at beginning of time step

delta_t: time step

#Output

azi_sp1: rotor azimuth (rad) at end of time step

Omega_sp1: rotor speed (Hz/s) at end of time step

OmegaDot_sp1: rotor acceleration (Hz/s) at end of time step

OWENS.userDefinedRotorSpeedProfile — Method

Internal, unused, userDefinedRotorSpeedProfile

OWENS.viz — Method

viz(;mesh=[],meshFile="none",resultsFile="none",selectedMode=10,sf=10)

Plots the mode shapes of a mode from a modal analysis performed using the OWENS toolkit. Either send the mesh struct or the mesh filename but not both

#Input

mesh::OWENSFEA.Mesh: Mesh struct, See ?OWENSFEA.Mesh

meshFile::String: optional string containing mesh file name

resultsFile::String: optional string containing results file name

selectedMode::Int: integer denoting selected mode to plot

sf::Float: scale factor for mode shape displacements

#Output None

OWENS.writeOwensNDL — Method

writeOwensNDL(fileRoot, nodes, cmkType, cmkValues)

writes a nodal input file

#Intput

fileRoot::string: string path to desired location with name but no extension

nodes::int: node numbers for C/M/K

cmkType::string: "C" "M" or "K"

cmkValues::float: C/M/K value

#Output

none: